refactor(protocols): split stack.go (1264 → 429 lines) by concern
Stack mixed lifecycle, packet-processing goroutines, SNMP integration,
and neighbor tracking in one file. Same-package split, no behaviour
or signature changes.

  stack.go (slim)      constants, types (Stack/Statistics/PacketObserver),
                       NewStack, Start/Stop, public API surface
                       (Send/Send Raw/GetDevices/ReloadConfig/GetStats,
                        debug-level accessors, currentConfig, capture filter)
  stack_init.go        AddPacketObserver, notifyObservers, initializeDevices,
                       resetDeviceState, register*, configureDHCPServer,
                       applySNMPAddrMappings
  stack_threads.go     receive/decode/route/send/babble goroutines + helpers
                       (routeByMAC, matchMAC, routeByEtherType, sendPacket)
  stack_snmp.go        initSNMPAgent, snmpEnabled, getSNMPAgents,
                       updateFDBTables, updateDot1DFdbTable, updateDot1QFdbTable
  stack_neighbors.go   recordNeighbor, neighbor cleanup loop,
                       selectDiscoveryDevice, isDeviceEnabledForProtocol

go test ./internal/protocols/... passes; golangci-lint clean.
